requirements.txt
fastapi
uvicorn
pydantic
playwright
python-dotenv
